moved ad list back in shared folder
