+++ src/app/shared/components/ads/ad-list/ad-list.component.scss
